requirements.txt
# Deep Learning Framework
torch>=2.0.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
tensorboard>=2.13.0

# Data Recording
mcap>=0.1.0
mcap-ros2-support>=0.1.0
